+++ NEWS.md
@@ -1,3 +1,7 @@
+* Added script to parse temperature data from CEDA into a data frame
+
+* Added working exergy-to-energy ratio (phi.u) scripts for graphs and reports.
+
 # SEAPSUTWorkflow 0.1.11 (2020-09-28)
 
 * Hotfix: fixed eta_fu_graph function